backend/pyproject.toml
[tool.poetry]
name = "educative-backend"
version = "0.1.0"
description = "Educative Project Backend (FastAPI)"
authors = ["Team <team@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.1"
pydantic = "^2.8.0"
SQLAlchemy = "^2.0.31"
httpx = "^0.27.0"
email-validator = "^2.3.0"
pynacl = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.5.6"
black = "^24.4.2"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
